more cleanup

+++ Content/Projectiles/Pets/MocoPet.cs
@@ -77,7 +77,7 @@ namespace Spooky.Content.Projectiles.Pets
                     num++;
                 }
             }
-            direction.X -= (float)((10 + num * 40) * player.direction);
+            direction.X -= (10 + num * 40) * player.direction;
             direction.Y -= 70f;
             float distanceTo = direction.Length();
             if (distanceTo > 200f && speed < 9f)
